Update schema, add rebuild
Thank you to github.com/Nenkai  for extensive reverse engineering as well as to github.com/DaZombieKiller for RE contributions!

This commit does not actually decode the half-float format (thought to be signed normalized floats) nor does it decompress DCT streams, and is mostly useful at this time for endian swapping.